requirements.txt
requests
Mastodon.py
